Autopause on timeout WIP

+++ packages/api/internal/orchestrator/cache.go
@@ -150,7 +150,6 @@ func (o *Orchestrator) getDeleteInstanceFunction(ctx context.Context, posthogCli
 			o.dns.Remove(ctx, info.Instance.SandboxID, node.Info.IPAddress)
 		}
 
-		req := &orchestrator.SandboxDeleteRequest{SandboxId: info.Instance.SandboxID}
 		if node == nil {
 			log.Printf("node '%s' not found", info.Instance.ClientID)
 			return fmt.Errorf("node '%s' not found", info.Instance.ClientID)
@@ -161,9 +160,17 @@ func (o *Orchestrator) getDeleteInstanceFunction(ctx context.Context, posthogCli
 			return fmt.Errorf("client for node '%s' not found", info.Instance.ClientID)
 		}
 
-		_, err = node.Client.Sandbox.Delete(ctx, req)
-		if err != nil {
-			return fmt.Errorf("failed to delete sandbox '%s': %w", info.Instance.SandboxID, err)
+		if info.AutoPause {
+			err = o.PauseInstance(ctx, &info, info.Instance.TemplateID, info.BuildID.String())
+			if err != nil {
+				return fmt.Errorf("failed to auto pause sandbox '%s': %w", info.Instance.SandboxID, err)
+			}
+		} else {
+			req := &orchestrator.SandboxDeleteRequest{SandboxId: info.Instance.SandboxID}
+			_, err = node.Client.Sandbox.Delete(ctx, req)
+			if err != nil {
+				return fmt.Errorf("failed to delete sandbox '%s': %w", info.Instance.SandboxID, err)
+			}
 		}
 
 		return nil
